move NavList to be its own folder under src
