+++ packages/components/src/NavList/NavList.story.tsx
@@ -26,31 +26,20 @@
 
 import React from 'react'
 import { Pivot } from '@looker/icons'
-import { renderWithTheme } from '@looker/components-test-utils'
-import { screen } from '@testing-library/react'
-import { ListItem } from './ListItem'
+import { DateRange } from '@styled-icons/material-outlined'
+import { ListItem } from '../List/ListItem'
 import { NavList } from './NavList'
 
-describe('NavList', () => {
-  test('display with keyColor', () => {
-    renderWithTheme(
-      <NavList>
-        <ListItem icon={<Pivot />} description="Orange-y" detail="Netherlands">
-          Explore
-        </ListItem>
-        <ListItem icon={<Pivot />}>Develop</ListItem>
-      </NavList>
-    )
+export default {
+  component: NavList,
+  title: 'NavList',
+}
 
-    const listItem = screen.getByText('Explore')
-
-    // NavList label
-    expect(listItem).toHaveStyle(`color: #6C43E0;`)
-
-    // NavList description
-    expect(screen.getByText('Orange-y')).toHaveStyle(`color: #6C43E0;`)
-
-    // NavList detail
-    expect(screen.getByText('Netherlands')).toHaveStyle(`color: #6C43E0;`)
-  })
-})
+export const navList = () => (
+  <NavList>
+    <ListItem icon={<Pivot />} description="Orange-y" detail="Netherlands">
+      Explore
+    </ListItem>
+    <ListItem icon={<DateRange />}>Develop</ListItem>
+  </NavList>
+)
